♻️ refactor:  テスト関連のフォルダ構成変更
Issue #145
