+++ test/domain/entities/owner_test.dart
@@ -6,7 +6,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:github_search/domain/entities/owner.dart';
 import 'package:github_search/infrastructure/github/json_object/owner/owner.dart';
 
-import '../../../test_utils/utils.dart';
+import '../../test_utils/utils.dart';
 
 void main() {
   final ownerJsonObject = TestAssets.readJsonMap('github/owner.json')!;
