+++ test/domain/repositories/app_data_repository_test.dart
@@ -8,9 +8,9 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:github_search/domain/repositories/app_data_repository.dart';
 
-import '../../../test_utils/hive.dart';
-import '../../../test_utils/locale.dart';
-import '../../../test_utils/mocks.dart';
+import '../../test_utils/hive.dart';
+import '../../test_utils/locale.dart';
+import '../../test_utils/mocks.dart';
 
 void main() {
   late Directory tmpDir;
