+++ test/domain/repositories/repo_repository_test.dart
@@ -6,8 +6,8 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:github_search/domain/repositories/repo_repository.dart';
 
-import '../../../test_utils/locale.dart';
-import '../../../test_utils/mocks.dart';
+import '../../test_utils/locale.dart';
+import '../../test_utils/mocks.dart';
 
 void main() {
   late ProviderContainer container;
