+++ test/presentation/common/pages/error_page_test.dart
@@ -6,8 +6,8 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:github_search/localizations/strings.g.dart';
 import 'package:github_search/presentation/common/pages/error_page.dart';
 
-import '../../test_utils/locale.dart';
-import '../../test_utils/mocks.dart';
+import '../../../test_utils/locale.dart';
+import '../../../test_utils/mocks.dart';
 
 void main() {
   setUp(useEnvironmentLocale);
