+++ test/presentation/repo/pages/index_page_test.dart
@@ -17,10 +17,10 @@ import 'package:github_search/presentation/repo/pages/index_page.dart';
 import 'package:github_search/presentation/repo/pages/search_page.dart';
 import 'package:github_search/presentation/repo/pages/view_page.dart';
 
-import '../../test_utils/hive.dart';
-import '../../test_utils/locale.dart';
-import '../../test_utils/logger.dart';
-import '../../test_utils/mocks.dart';
+import '../../../test_utils/hive.dart';
+import '../../../test_utils/locale.dart';
+import '../../../test_utils/logger.dart';
+import '../../../test_utils/mocks.dart';
 
 void main() {
   late Directory tmpDir;
